Moar compression settings.

+++ src/internal/blosc.compression.params.cpp
@@ -1,4 +1,4 @@
-#include "blosc.compressor.hh"
+#include "blosc.compression.params.hh"
 
 zarr::BloscCompressionParams::BloscCompressionParams()
   : clevel{ 1 }
@@ -7,8 +7,8 @@ zarr::BloscCompressionParams::BloscCompressionParams()
 }
 
 zarr::BloscCompressionParams::BloscCompressionParams(std::string_view codec_id,
-                                                     int clevel,
-                                                     int shuffle)
+                                                     uint8_t clevel,
+                                                     uint8_t shuffle)
   : codec_id{ codec_id }
   , clevel{ clevel }
   , shuffle{ shuffle }
